requirements.txt
django
djangorestframework
#markdown
#django-filter